Restore fix from commit 2e324352d (fixes tracker 3102439)

+++ src/core/CLucene/index/IndexWriter.cpp
@@ -99,7 +99,7 @@ void IndexWriter::ensureOpen()   {
 void IndexWriter::message(string message) {
   if (infoStream != NULL){
     (*infoStream) << string("IW ") << Misc::toString(messageID) << string(" [")
-    						  << Misc::toString( (int32_t)(_LUCENE_CURRTHREADID) ) << string("]: ") << message << string("\n");
+    						  << Misc::toString( _LUCENE_CURRTHREADID ) << string("]: ") << message << string("\n");
   }
 }
